+++ src/dayjs/max-duration/index.ts
@@ -1,0 +1,2 @@
+export { MaxDuration, MAX_DURATION } from './max-duration.decorator'
+export { maxDuration } from './max-duration.predicate'