+++ src/dayjs/min-duration/index.ts
@@ -1,0 +1,2 @@
+export { MinDuration, MIN_DURATION } from './min-duration.decorator'
+export { minDuration } from './min-duration.predicate'